pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "refine-motion"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "websockets>=12",
    "anthropic>=0.28",
    "httpx>=0.27",
    # STT
    "openai-whisper>=20231117",
    # audio i/o on server side
    "soundfile>=0.12",
    "numpy>=1.26",
    # codebase traversal
    "gitpython>=3.1",
    "pathspec>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx",   # TestClient
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
